Extract shared test utilities from duplicated mock boilerplate
Create InMemoryCache in Caching for test isolation, PlaylistTesting library with MockPlaylistFetcher and PlaylistStubs, and MockFeatureFlagProvider in AnalyticsTesting. Refactor 22 test files across 6 packages to use shared utilities, eliminating 537 lines of duplicated mock code.
