+++ Shared/Metadata/Tests/MetadataTests/PlaycutMetadataServiceCachingTests.swift
@@ -12,6 +12,7 @@ import Testing
 import Foundation
 import Core
 import Playlist
+import PlaylistTesting
 @testable import Caching
 @testable import Metadata
 
@@ -98,33 +99,6 @@ final class MetadataMockWebSession: WebSession, @unchecked Sendable {
         responses.removeAll()
         requestedURLs.removeAll()
         requestCount = 0
-    }
-}
-
-// MARK: - Playcut Test Stub
-
-extension Playcut {
-    /// Creates a Playcut with sensible defaults for testing.
-    static func stub(
-        id: UInt64 = 1,
-        hour: UInt64 = 1000,
-        chronOrderID: UInt64? = nil,
-        timeCreated: UInt64? = nil,
-        songTitle: String = "Test Song",
-        labelName: String? = nil,
-        artistName: String = "Test Artist",
-        releaseTitle: String? = "Test Album"
-    ) -> Playcut {
-        Playcut(
-            id: id,
-            hour: hour,
-            chronOrderID: chronOrderID ?? id,
-            timeCreated: timeCreated ?? hour,
-            songTitle: songTitle,
-            labelName: labelName,
-            artistName: artistName,
-            releaseTitle: releaseTitle
-        )
     }
 }
