+++ Shared/Playlist/Package.swift
@@ -4,7 +4,10 @@ import PackageDescription
 let package = Package(
     name: "Playlist",
     platforms: [.iOS("18.4"), .watchOS(.v11), .macOS(.v15)],
-    products: [.library(name: "Playlist", targets: ["Playlist"])],
+    products: [
+        .library(name: "Playlist", targets: ["Playlist"]),
+        .library(name: "PlaylistTesting", targets: ["PlaylistTesting"]),
+    ],
     dependencies: [
         .package(name: "Analytics", path: "../Analytics"),
         .package(name: "Core", path: "../Core"),
@@ -17,10 +20,15 @@ let package = Package(
             dependencies: ["Analytics", "Core", "Caching", "Logger"],
             resources: [.process("Playlist Detail Assets.xcassets")]
         ),
+        .target(
+            name: "PlaylistTesting",
+            dependencies: ["Playlist"]
+        ),
         .testTarget(
             name: "PlaylistTests",
             dependencies: [
                 "Playlist",
+                "PlaylistTesting",
                 "Caching",
                 .product(name: "AnalyticsTesting", package: "Analytics"),
                 .product(name: "LoggerTesting", package: "Logger"),
