+++ Shared/Playlist/Tests/PlaylistTests/PlaylistServiceTests.swift
@@ -10,82 +10,15 @@
 
 import Testing
 import Foundation
+import PlaylistTesting
 @testable import Playlist
 @testable import Caching
 
-// MARK: - Mock PlaylistFetcher
+// MARK: - Helpers
 
-final class MockPlaylistFetcher: PlaylistFetcherProtocol, @unchecked Sendable {
-    var playlistToReturn: Playlist = .empty
-    var callCount = 0
-
-    func fetchPlaylist() async -> Playlist {
-        callCount += 1
-        return playlistToReturn
-    }
-}
-
-// MARK: - Mock Cache for Tests
-
-/// In-memory cache for isolated test execution
-final class PlaylistTestMockCache: Cache, @unchecked Sendable {
-    private var dataStorage: [String: Data] = [:]
-    private var metadataStorage: [String: CacheMetadata] = [:]
-    private let lock = NSLock()
-
-    func metadata(for key: String) -> CacheMetadata? {
-        lock.lock()
-        defer { lock.unlock() }
-        return metadataStorage[key]
-    }
-    
-    func data(for key: String) -> Data? {
-        lock.lock()
-        defer { lock.unlock() }
-        return dataStorage[key]
-    }
-
-    func set(_ data: Data?, metadata: CacheMetadata, for key: String) {
-        lock.lock()
-        defer { lock.unlock() }
-        if let data = data {
-            dataStorage[key] = data
-            metadataStorage[key] = metadata
-        } else {
-            remove(for: key)
-        }
-    }
-    
-    func remove(for key: String) {
-        lock.lock()
-        defer { lock.unlock() }
-        dataStorage.removeValue(forKey: key)
-        metadataStorage.removeValue(forKey: key)
-    }
-
-    func allMetadata() -> [(key: String, metadata: CacheMetadata)] {
-        lock.lock()
-        defer { lock.unlock() }
-        return metadataStorage.map { ($0.key, $0.value) }
-    }
-
-    func clearAll() {
-        lock.lock()
-        defer { lock.unlock() }
-        dataStorage.removeAll()
-        metadataStorage.removeAll()
-    }
-
-    func totalSize() -> Int64 {
-        lock.lock()
-        defer { lock.unlock() }
-        return dataStorage.values.reduce(0) { $0 + Int64($1.count) }
-    }
-}
-
-/// Helper to create an isolated cache coordinator for testing
+/// Helper to create an isolated cache coordinator for testing.
 func makeTestCacheCoordinator() -> CacheCoordinator {
-    CacheCoordinator(cache: PlaylistTestMockCache())
+    CacheCoordinator(cache: InMemoryCache())
 }
 
 // MARK: - Tests
@@ -384,7 +317,7 @@ struct PlaylistServiceTests {
     @Test("Observer receives cached data even if subscription happens during cache load", .timeLimit(.minutes(1)))
     func observerReceivesCachedDataDuringCacheLoad() async throws {
         // Given - Pre-populate cache before creating service
-        let mockCache = PlaylistTestMockCache()
+        let mockCache = InMemoryCache()
         let cacheCoordinator = CacheCoordinator(cache: mockCache)
         let cachedPlaylist = Playlist.stub(playcuts: [.stub(songTitle: "Pre-cached Song", artistName: "Pre-cached Artist")])
